+++ src/mako/security/password/HasherException.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\security\password;
+
+use RuntimeException;
+
+/**
+ * Hasher exception.
+ *
+ * @author Frederic G. Østby
+ */
+class HasherException extends RuntimeException
+{
+
+}